Beam distribution within a sample
Beam distribution within a sample. A) Widefield microscopy – the beam is illuminating the entire sample above and below the plane of focus. B) A confocal beam is illuminating the sample only partially but needs to be scanned C) and is of a higher power. D) The theoretical layout of a light sheet system, where the excitation beam is placed perpendicular to the imaging axis. (Adapted from Brad Amos, Cambridge)

Original image: https://www.photometrics.com/wp-content/uploads/2019/12/What-Is-Light-Sheet_image1-1024x580.png