confocal vs light sheet microscopy
Light-sheet microscopy is a fluorescence imaging technique, which utilizes a sheet of laser light to illuminate only a thin slice of the sample.

The basic technical principle is a wide-field fluorescence microscope, placed perpendicular to the light-sheet, that collects the fluorescence signal and images of the observed region by means of a full-frame camera. The orthogonal arrangement that decouples the illumination from the detection enables intrinsic 3D optical sectioning, as compared to other fluorescent imaging techniques like confocal and spinning disc microscopy. As a result, the method features drastically reduced overall acquisition duration, photobleaching effects and phototoxicity, as well as yields excellent signal-to-noise ratio and enables high temporal and 3D-spatial resolution.

Light-sheet fluorescence microscopy (LSFM) can be utilized to image a huge variety of fixed, live or cleared biological samples. Applications of light-sheet microscopy can range from imaging of subcellular structures and rapid inter- and intracellular processes to the acquisition of the long-term development of a model system, to the complete visualization of a macroscale cleared sample.

Original image: https://www.bruker.com/en/products-and-solutions/fluorescence-microscopy/light-sheet-microscopes/what-is-light-sheet-microscopy/_jcr_content/root/contentpar/twocolumns_copy/contentpar-2/image.coreimg.82.1280.png/1600244164830/comparison-of-confocal-microscopy-and-light-sheet-microscopy-b-bruker-jpg.png