iSPIM
The system is built on a vertical breadboard mounted on top of a standard optical table. The lasers (a 405 nm Coherent Obis solid state diode laser and a Chameleon Ultra 2 to provide two-photon excitation) send the light to two separate beam shaping units to provide comparable dimension of the beams. Mirrors and a periscope bring the laser beams to the vertical breadboard. The cylindrical lens creates the light sheets, which are again focused on the sample by the illumination objective. The detection objective collects the fluorescence and sends it to a tube lens, which focuses the light onto the chip of the SCMOS camera, while a dichroic mirror and a fluorescence filter select the correct wavelength for the detected light.

source: https://www.nature.com/articles/srep23923/figures/1